\name{expdata}
\alias{expdata}

\title{
Real E.coli Gene Expression Dataset
}
\description{
Normalized real microarray gene expression E.coli data set.
}
\usage{data("expdata")}

\details{
This dataset compiled in (Faith, 2007) . The example data set is also given to provide a sample data set so that one can take it as reference while formatting the new data set as entry to  As in this example datset, the input data sets must be complete and neither empty or non-numerical characters are allowed. 
}
\source{
This data was downloaded from m3d.mssm.edu/network_inference.html
}
\references{
Faith JJ, Hayete B, Thaden JT, Mogno I, Wierzbowski J, Cottarel G, Kasif S, Collins JJ, Gardner TS. Large-scale mapping and validation of Escherichia coli transcriptional regulation from a compendium of expression profiles. PLoS Biol. 2007 Jan;5(1):e8
}
\examples{

data(expdata) #returns data matrix with the name expdata.

}
\keyword{datasets}
